+++ src/package/themes/default/index.ts
@@ -16,6 +16,7 @@ import popover from './components/popover';
 import dropdown from './components/dropdown';
 import splitbutton from './components/splitbutton';
 import contextMenu from './components/context-menu';
+import commandPalette from './components/command-palette';
 import select from './components/select';
 import autocomplete from './components/autocomplete';
 import multiselect from './components/multiselect';
@@ -64,6 +65,7 @@ export default {
         dropdown,
         splitbutton,
         contextMenu,
+        commandPalette,
         select,
         autocomplete,
         multiselect,